refactor: трансформация полученных через API данных перенесена в BurgerIngridients
